Fixed issue in reports.py where decimal fees would cause an error

+++ app/reports.py
@@ -44,7 +44,7 @@ def countTimes(column, start, end):
 
 def convertToCurrency(x):
     if x != None:
-        return "£{:,.2f}".format(int(x))
+        return "£{:,.2f}".format(float(x))
 
     return None
